+++ x/gov/legacy/v220/types.go
@@ -1,0 +1,4 @@
+package v220
+
+type Deposits []Deposit
+type Votes []Vote