+++ CHANGELOG.md
@@ -43,6 +43,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### API Breaking Changes
 ### State Machine Breaking Changes
 * (app) [\#326](https://github.com/certikfoundation/shentu/pull/326) Bump Cosmos SDK to v0.44.3.
+* (app) [\#334](https://github.com/certikfoundation/shentu/pull/334) Implement in-store migration from v2.2.0 to v2.3.0.
+* (x/gov) [\#334](https://github.com/certikfoundation/shentu/pull/334) `TxHash` field has been removed from `Vote` and `Deposit` types.
 
 ### Features
 * (app) [\#326](https://github.com/certikfoundation/shentu/pull/326) Add `authz` and `feegrant` modules.
@@ -55,7 +57,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * (x/shield) [\#326](https://github.com/certikfoundation/shentu/pull/326) Add checks for expired entries in shield purchase.
 * (x/gov) [\#331](https://github.com/certikfoundation/shentu/pull/331) Fix gov tally logic.
 
-## [v2.1.0] - 08-09-2021
+
+## [v2.1.0] - 09-03-2021
 
 Version 2.1.0 re-enables endblocker in the staking module, and bumps SDK to 0.42.9 for necessary query route.